TP3-V5
Game now work. Player created.

-Add "Button" class for the button menu (menu is not implemented yet)

-Add "State" for managing all states

-Add "GameplayState" for initializing everything

-Add player to the screen with movement and restrictions

+++ TP3/Ejercicio-1/Object.cpp
@@ -1,5 +1,5 @@
 #include "Object.h"
-Object::Object(float x, float y, const char* imagePath) {
+Object::Object(int x, int y, const char* imagePath) {
 	_x = x;
 	_y = y;
 	_image = al_load_bitmap(imagePath);
@@ -7,10 +7,10 @@ Object::Object(float x, float y, const char* imagePath) {
 Object::~Object() {
 	al_destroy_bitmap(_image);
 }
-float Object::getX() {
+int Object::getX() {
 	return _x;
 }
-float Object::getY() {
+int Object::getY() {
 	return _y;
 }
 ALLEGRO_BITMAP* Object::getImage() {
